Ajout de la documentation pour Swagger dans l'API Fichiers.
